Update PilotsList to be connected
Connected the PilotsList and PilotsListRow components
PilotsList now extracts a list of pilot IDs instead of objects
Disconnected Pilots, which is now effectively presentational again
