+++ src/features/pilots/PilotsList/PilotsListRow.jsx
@@ -1,6 +1,41 @@
 import React from "react";
+import {connect} from "react-redux";
 import {Table} from "semantic-ui-react";
 import _ from "lodash";
+
+import schema from "app/schema";
+
+const mapState = (state, ownProps) => {
+    const session = schema.from(state.entities);
+    const {Pilot} = session;
+
+    let pilot;
+
+    if(Pilot.hasId(ownProps.pilotID)) {
+        const pilotModel = Pilot.withId(ownProps.pilotID);
+
+        // Access the underlying plain JS object using the "ref" field,
+        // and make a shallow copy of it
+        pilot = {
+            ...pilotModel.ref
+        };
+
+        // We want to look up pilotModel.mech.mechType.  Just in case the
+        // relational fields are null, we'll do a couple safety checks as we go.
+
+        // Look up the associated Mech instance using the foreign-key
+        // field that we defined in the Pilot Model class
+        const {mech} = pilotModel;
+
+        // If there actually is an associated mech, include the
+        // mech type's ID as a field in the data passed to the component
+        if(mech && mech.type) {
+            pilot.mechType = mech.type.id;
+        }
+    }
+
+    return {pilot};
+}
 
 const PilotsListRow = ({pilot={}, onPilotClicked=_.noop, selected}) => {
     const {
@@ -34,4 +69,4 @@ const PilotsListRow = ({pilot={}, onPilotClicked=_.noop, selected}) => {
     );
 }
 
-export default PilotsListRow;
+export default connect(mapState)(PilotsListRow);
